pyproject.toml
[project]
name = "at-bus-infrastructure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ansible>=11.7.0",
    "apache-airflow==2.10.5",
    "apache-airflow-providers-docker>=4.4.0",
    "apache-airflow-providers-google>=15.1.0",
    "astronomer-cosmos>=1.10.1",
    "dbt-bigquery>=1.9.2",
    "dbt-core>=1.9.8",
    "polars>=1.30.0",
    "ruff>=0.11.11",
]
